+++ packages/base/core/__tests__/Collection.bench.ts
@@ -53,38 +53,64 @@ describe('Collection benchmarks', () => {
       col2.findOne({ num: 999 })
     })
   })
-  describe('transformAll', () => {
-    const col1 = new Collection()
-    interface TestItem {
-      id: number,
-      parent?: any,
-    }
-    const transformAll: TransformAll<BaseItem, TestItem> = (items, fields) => {
-      if (fields?.parent) {
-        const foreignKeys = [...new Set(items.map(item => item.parent))]
-        const relatedItems = col1.find({ id: { $in: foreignKeys } }).fetch()
-        items.forEach((item) => {
-          item.parent = relatedItems.find(related => related.id === item.parent)
+    describe('index null and undefined values', () => {
+        const col1 = new Collection<{ id: string, name: string, num?: number | null }>({
+            indices: [createIndex('num')],
         })
-      }
-      return items
-    }
-    const col2 = new Collection({ transformAll })
+        const col2 = new Collection<{ id: string, name: string, num?: number | null }>()
 
-    Collection.batch(() => {
-      // create items
-      for (let i = 0; i < 10_000; i += 1) {
-        col1.insert({ id: i.toString(), name: 'John Sr.', num: i })
-        col2.insert({ id: i.toString(), name: 'John', parent: i.toString() })
-      }
-    })
+        Collection.batch(() => {
+            // create items
+            for (let i = 0; i < 10_000; i += 1) {
+                col1.insert({ id: i.toString(), name: 'John', num: i > 5000 ? i : undefined })
+                col2.insert({ id: i.toString(), name: 'John', num: i > 5000 ? i : undefined })
+            }
+        })
 
-    bench('default', () => {
-      col2.find().map(value => value.parent = col1.findOne({ id: value.parent }))
-    })
+        bench('with index', () => {
+            col1.findOne({ num: undefined })
+            col1.findOne({ num: null })
+            col1.findOne({ num: { $exists: false } })
+        })
 
-    bench('transformAll', () => {
-      col2.find({}, { fields: { parent: 1 } }).fetch()
+        bench('without index', () => {
+            col2.findOne({ num: undefined })
+            col2.findOne({ num: null })
+            col2.findOne({ num: { $exists: false } })
+        })
     })
-  })
+    describe('transformAll', () => {
+        const col1 = new Collection()
+        interface TestItem {
+            id: number,
+            parent?: any,
+        }
+        const transformAll: TransformAll<BaseItem, TestItem> = (items, fields) => {
+            if (fields?.parent) {
+                const foreignKeys = [...new Set(items.map(item => item.parent))]
+                const relatedItems = col1.find({ id: { $in: foreignKeys } }).fetch()
+                items.forEach((item) => {
+                    item.parent = relatedItems.find(related => related.id === item.parent)
+                })
+            }
+            return items
+        }
+        const col2 = new Collection({ transformAll })
+
+        Collection.batch(() => {
+            // create items
+            for (let i = 0; i < 10_000; i += 1) {
+                col1.insert({ id: i.toString(), name: 'John Sr.', num: i })
+                col2.insert({ id: i.toString(), name: 'John', parent: i.toString() })
+            }
+        })
+
+        bench('default', () => {
+            col2.find().map(value => value.parent = col1.findOne({ id: value.parent }))
+        })
+
+        bench('transformAll', () => {
+            col2.find({}, { fields: { parent: 1 } }).fetch()
+        })
+    })
 })