+++ packages/devtools/devtools/src/utils/useCollectionMutations.ts
@@ -12,8 +12,8 @@ export default function useCollectionMutations(collectionName: string) {
     if (!mutations) return []
     return mutations.items
       .filter(q => q.collectionName === collectionName)
-      .sort((a, b) => a.time - b.time)
-      .reverse()
+      .toSorted((a, b) => a.time - b.time)
+      .toReversed()
       .map(({ collectionName: _, time, ...item }) => ({
         id: item.id,
         time: new Date(time as number),