+++ packages/devtools/devtools/src/utils/useCollectionQueries.ts
@@ -12,8 +12,8 @@ export default function useCollectionQueries(collectionName: string) {
     if (!queries) return []
     return queries.items
       .filter(q => q.collectionName === collectionName)
-      .sort((a, b) => a.time - b.time)
-      .reverse()
+      .toSorted((a, b) => a.time - b.time)
+      .toReversed()
       .map(({ collectionName: _, lastTime, ...item }) => ({
         id: item.id,
         collectionName: item.collectionName,